# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GreenBecak API",
    "description": "API lengkap untuk aplikasi manajemen becak GreenBecak - Updated v2.0",
    "version": "2.0.0",
    "contact": {
      "name": "GreenBecak Team",
      "email": "support@greenbecak.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token dalam format: Bearer <token>"
    }
  },
  "paths": {
...
    "/notifications/{id}": {
      "get": {
        "summary": "Get Notification by ID",
        "description": "Mendapatkan detail notifikasi berdasarkan ID",
        "tags": ["Notifications"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Detail notifikasi",
            "schema": {
              "type": "object",
              "properties": {
                "notification": {"$ref": "#/definitions/Notification"}
              }
            }
          },
          "404": {
            "description": "Notifikasi tidak ditemukan"
          }
        }
      },
      "put": {
        "summary": "Mark Notification as Read",
        "description": "Tandai notifikasi sebagai sudah dibaca",
        "tags": ["Notifications"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifikasi berhasil ditandai sebagai dibaca",
            "schema": {
              "type": "object",
              "properties": {
                "message": {"type": "string"},
                "notification": {"$ref": "#/definitions/Notification"}
              }
            }
          },
          "404": {
            "description": "Notifikasi tidak ditemukan"
          }
        }
      },
      "delete": {
        "summary": "Delete Notification",
        "description": "Hapus notifikasi",
        "tags": ["Notifications"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifikasi berhasil dihapus",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Notification deleted successfully"
                }
              }
            }
          },
          "404": {
            "description": "Notifikasi tidak ditemukan"
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {"type": "integer"},
        "username": {"type": "string"},
        "email": {"type": "string"},
        "name": {"type": "string"},
        "role": {"type": "string", "enum": ["admin", "customer"]},
        "phone": {"type": "string"},
        "address": {"type": "string"},
        "is_active": {"type": "boolean"},
        "created_at": {"type": "string", "format": "date-time"},
        "updated_at": {"type": "string", "format": "date-time"}
      }
    },
...
    "Notification": {
      "type": "object",
      "properties": {
        "id": {"type": "integer"},
        "user_id": {"type": "integer"},
        "title": {"type": "string"},
        "message": {"type": "string"},
        "type": {"type": "string", "enum": ["info", "success", "warning", "error"]},
        "is_read": {"type": "boolean"},
        "created_at": {"type": "string", "format": "date-time"},
        "updated_at": {"type": "string", "format": "date-time"},
        "user": {"$ref": "#/definitions/User"}
      }
    },
...
  },
...
}